[781] Rename plaid_transaction table to raw_plaid_transactions
