+++ app/jobs/raw_plaid_transaction_job/imports.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-module PlaidTransactionJob
+module RawPlaidTransactionJob
   class Imports < ApplicationJob
     def perform
       BankAccount.syncing.pluck(:id).each do |bank_account_id|
-        ::PlaidTransactionService::Plaid::Import.new(bank_account_id: bank_account_id).run
+        ::RawPlaidTransactionService::Plaid::Import.new(bank_account_id: bank_account_id).run
       end
     end
   end
