--- app/models/plaid_transaction.rb
@@ -1,3 +1,0 @@
-class PlaidTransaction < ApplicationRecord
-  monetize :amount_cents
-end
