+++ app/models/raw_plaid_transaction.rb
@@ -1,0 +1,3 @@
+class RawPlaidTransaction < ApplicationRecord
+  monetize :amount_cents
+end
