--- app/services/plaid_transaction_service/likely/ach.rb
@@ -1,9 +1,0 @@
-module PlaidTransactionService
-  module Likely
-    class Ach
-      def run
-        ::PlaidTransaction.where("plaid_transaction->>'name' ilike '%busbillpay%'")
-      end
-    end
-  end
-end
