--- app/services/plaid_transaction_service/likely/disbursement.rb
@@ -1,9 +1,0 @@
-module PlaidTransactionService
-  module Likely
-    class Disbursement
-      def run
-        ::PlaidTransaction.where("plaid_transaction->>'name' ilike '%hcb disburse%'")
-      end
-    end
-  end
-end
