--- app/services/plaid_transaction_service/likely/emburse.rb
@@ -1,9 +1,0 @@
-module PlaidTransactionService
-  module Likely
-    class Emburse
-      def run
-        ::PlaidTransaction.where("plaid_transaction->>'name' ilike '%emburse%' or plaid_transaction->>'merchant' ilike '%emburse%'")
-      end
-    end
-  end
-end
