--- app/services/plaid_transaction_service/likely/expensify.rb
@@ -1,9 +1,0 @@
-module PlaidTransactionService
-  module Likely
-    class Expensify
-      def run
-        ::PlaidTransaction.where("plaid_transaction->>'name' ilike '%expensify%'")
-      end
-    end
-  end
-end
