--- app/services/plaid_transaction_service/likely/fee_payment.rb
@@ -1,9 +1,0 @@
-module PlaidTransactionService
-  module Likely
-    class FeePayment
-      def run
-        ::PlaidTransaction.where("plaid_transaction->>'name' ilike '%bank fee%'")
-      end
-    end
-  end
-end
