--- app/services/plaid_transaction_service/likely/fee_reimbursement.rb
@@ -1,9 +1,0 @@
-module PlaidTransactionService
-  module Likely
-    class FeeReimbursement
-      def run
-        ::PlaidTransaction.where("plaid_transaction->>'name' ilike '%fee refund%'")
-      end
-    end
-  end
-end
