--- app/services/plaid_transaction_service/likely/github.rb
@@ -1,9 +1,0 @@
-module PlaidTransactionService
-  module Likely
-    class Github
-      def run
-        ::PlaidTransaction.where("plaid_transaction->>'name' ilike '%github grant%'")
-      end
-    end
-  end
-end
