+++ app/services/raw_plaid_transaction_service/likely/ach.rb
@@ -1,0 +1,9 @@
+module RawPlaidTransactionService
+  module Likely
+    class Ach
+      def run
+        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%busbillpay%'")
+      end
+    end
+  end
+end
