+++ app/services/raw_plaid_transaction_service/likely/expensify.rb
@@ -1,0 +1,9 @@
+module RawPlaidTransactionService
+  module Likely
+    class Expensify
+      def run
+        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%expensify%'")
+      end
+    end
+  end
+end
