+++ app/services/raw_plaid_transaction_service/likely/fee_payment.rb
@@ -1,0 +1,9 @@
+module RawPlaidTransactionService
+  module Likely
+    class FeePayment
+      def run
+        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%bank fee%'")
+      end
+    end
+  end
+end
