+++ app/services/raw_plaid_transaction_service/likely/fee_reimbursement.rb
@@ -1,0 +1,9 @@
+module RawPlaidTransactionService
+  module Likely
+    class FeeReimbursement
+      def run
+        ::RawPlaidTransaction.where("plaid_transaction->>'name' ilike '%fee refund%'")
+      end
+    end
+  end
+end
