+++ db/migrate/20200915231730_rename_plaid_transactions_to_raw_plaid_transactions.rb
@@ -1,0 +1,7 @@
+class RenamePlaidTransactionsToRawPlaidTransactions < ActiveRecord::Migration[6.0]
+  def change
+    safety_assured do
+      rename_table :plaid_transactions, :raw_plaid_transactions
+    end
+  end
+end
